requirements.txt
evernote
python-slugify
jinja2
colorama
